fixup: add unit test & other comments
